feat: Balance 불러오기 api 수정

+++ src/main/java/io/heachi/backend/infra/blockchain/base/Blockchain.java
@@ -2,13 +2,12 @@ package io.heachi.backend.infra.blockchain.base;
 
 import java.io.IOException;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 
 public interface Blockchain {
 
   WalletInfo createWallet();
 
-  BigInteger getBalance(String address) throws IOException;
+  BigDecimal getBalance(String address) throws IOException;
 
   String transfer(String address, String privateKey, String toHash, BigDecimal eth);
 }
